Separate statedb initialization from blockchain initialization
